Refactor into connection, session start, launch
