+++ src/kernels/jupyter/connection/jupyterPasswordConnect.ts
