+++ src/kernels/jupyter/connection/preferredRemoteKernelIdProvider.ts
@@ -5,10 +5,10 @@ import { inject, injectable, named } from 'inversify';
 /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-require-imports */
 import cloneDeep from 'lodash/cloneDeep';
 import { Memento, Uri } from 'vscode';
-import { traceVerbose } from '../../platform/logging';
-import { getDisplayPath } from '../../platform/common/platform/fs-paths';
-import { IMemento, GLOBAL_MEMENTO, ICryptoUtils } from '../../platform/common/types';
-import { sendTelemetryEvent, Telemetry } from '../../telemetry';
+import { traceVerbose } from '../../../platform/logging';
+import { getDisplayPath } from '../../../platform/common/platform/fs-paths';
+import { IMemento, GLOBAL_MEMENTO, ICryptoUtils } from '../../../platform/common/types';
+import { sendTelemetryEvent, Telemetry } from '../../../telemetry';
 
 export const ActiveKernelIdList = 'Active_Kernel_Id_List';
 // This is the number of kernel ids that will be remembered between opening and closing VS code
