+++ src/kernels/jupyter/serviceRegistry.web.ts
@@ -6,20 +6,20 @@ import { IServiceManager } from '../../platform/ioc/types';
 import { DataScienceErrorHandlerWeb } from '../errors/kernelErrorHandler.web';
 import { IDataScienceErrorHandler } from '../errors/types';
 import { INotebookProvider } from '../types';
-import { JupyterConnection } from './jupyterConnection';
-import { JupyterKernelService } from './jupyterKernelService.web';
-import { JupyterRemoteCachedKernelValidator } from './jupyterRemoteCachedKernelValidator';
-import { JupyterUriProviderRegistration } from './jupyterUriProviderRegistration';
+import { JupyterConnection } from './connection/jupyterConnection';
+import { JupyterKernelService } from './session/jupyterKernelService.web';
+import { JupyterRemoteCachedKernelValidator } from './connection/jupyterRemoteCachedKernelValidator';
+import { JupyterUriProviderRegistration } from './connection/jupyterUriProviderRegistration';
 import { JupyterCommandLineSelector } from './launcher/commandLineSelector';
 import { JupyterNotebookProvider } from './launcher/jupyterNotebookProvider';
-import { JupyterPasswordConnect } from './launcher/jupyterPasswordConnect';
+import { JupyterPasswordConnect } from './connection/jupyterPasswordConnect';
 import { HostJupyterExecution } from './launcher/liveshare/hostJupyterExecution';
 import { HostJupyterServerFactory } from './launcher/liveshare/hostJupyterServerFactory';
 import { NotebookProvider } from './launcher/notebookProvider';
 import { NotebookServerProvider } from './launcher/notebookServerProvider';
-import { JupyterServerUriStorage } from './launcher/serverUriStorage';
-import { LiveRemoteKernelConnectionUsageTracker } from './liveRemoteKernelConnectionTracker';
-import { JupyterServerSelector } from './serverSelector';
+import { JupyterServerUriStorage } from './connection/serverUriStorage';
+import { LiveRemoteKernelConnectionUsageTracker } from './connection/liveRemoteKernelConnectionTracker';
+import { JupyterServerSelector } from './connection/serverSelector';
 import { BackingFileCreator } from './session/backingFileCreator.web';
 import { JupyterRequestCreator } from './session/jupyterRequestCreator.web';
 import { JupyterSessionManagerFactory } from './session/jupyterSessionManagerFactory';
